+++ ChangeLog.md
@@ -1,3 +1,4 @@
 # Changelog for HaskellGame
 
-Working road (lines) retro scrolling
+0.1.1.0 - Added working road (lines) retro scrolling
+0.1.3.0 - Added infinite racing track